Adding full page for selecting device to register
