--- frontend/src/pages/UserLogPage/index.ts
@@ -1,1 +1,0 @@
-export { UserLogPage } from './UserLogPage'
